+++ test/integration/bind/controls/default_spec.rb
@@ -1,10 +1,32 @@
 domain = input('domain')
 host_string = input('host_string')
+chroot = input('chroot')
+
+case os.family
+when 'debian'
+  chroot_cmd = '-t /var/bind9/chroot'
+  named_cmd = if os.name == 'debian' && os.release.to_i == 10
+                '/usr/sbin/named -u bind'
+              else
+                '/usr/sbin/named -f -u bind'
+              end
+when 'redhat', 'fedora'
+  named_cmd = '/usr/sbin/named -u named -c /etc/named.conf'
+  chroot_cmd = '-t /var/named/chroot'
+end
 
 control 'default' do
   describe port 53 do
     it { should be_listening }
     its('processes') { should include 'named' }
+  end
+
+  describe processes 'named' do
+    if chroot
+      its('commands') { should include "#{named_cmd} #{chroot_cmd}" }
+    else
+      its('commands') { should include named_cmd }
+    end
   end
 
   describe command 'rndc reload' do